+++ train_telegram.py
@@ -3,16 +3,13 @@ from tqdm import tqdm, trange
 import argparse
 from transformers import (
     AdamW,
-    GPT2LMHeadModel,
     GPT2Tokenizer,
-    OpenAIGPTLMHeadModel,
-    OpenAIGPTTokenizer,
-    get_linear_schedule_with_warmup, AutoModelWithLMHead, AutoTokenizer
+    get_linear_schedule_with_warmup, AutoModelWithLMHead,
 )
 
 from dataset.utils import set_seed, add_special_tokens_
 from dataset.text_dataset import get_data_loader
-from utils.sample import sample_sequence
+from utils.sample import sample, message_sample
 import logging
 import tensorboardX
 import os
@@ -22,118 +19,7 @@ logger = logging.getLogger(__name__)
 logging.basicConfig(level=logging.DEBUG, format='%(asctime)s [%(levelname)-5.5s] [%(name)-12.12s]: %(message)s')
 
 
-def parse_args():
-    parser = argparse.ArgumentParser(description='Training params')
-    parser.add_argument('--run_name', default='telegram', type=str)
-    parser.add_argument('--input_file', default='chat.txt', type=str)
-    parser.add_argument('--sample_file', default='chat.txt', type=str)
-    parser.add_argument('--sampling_history', default=None, type=str)
-    parser.add_argument('--val_file', default=None, type=str)
-    parser.add_argument('--model_type', default='gpt2', type=str)
-    parser.add_argument('--save_every', default=30, type=int)
-    parser.add_argument('--max_input_lenth', default=400, type=int)
-    parser.add_argument('--weight_decay', default=0, type=float)
-    parser.add_argument('--train_batch_size', default=2, type=int)
-    parser.add_argument('--gradient_accumulation_steps', default=16, type=int)
-    parser.add_argument('--warmup_steps', default=8, type=int)
-    parser.add_argument('--lr', default=5e-5, type=float)
-    parser.add_argument('--adam_epsilon', default=1e-8, type=float)
-    parser.add_argument('--max_norm', default=1, type=float)
-    parser.add_argument('--n_epochs', default=12, type=int)
-    parser.add_argument('--device', default='cuda', type=str)
-    parser.add_argument('--seed', default=42, type=int)
-    parser.add_argument('--always_save', default=1, type=int)
-    return parser.parse_args()
-
-
-def sample(model, tokenizer, sampling_history=None, max_history=2, no_info=True, max_generation_steps=10):
-    speaker1_tag = '<speaker1>'
-    speaker2_tag = '<speaker2>'
-    speaker1_tag_id = tokenizer.convert_tokens_to_ids(speaker1_tag)
-    speaker2_tag_id = tokenizer.convert_tokens_to_ids(speaker2_tag)
-    # history = f"""
-    # {speaker2_tag} Привет )
-    # {speaker1_tag} Привет, чё как?
-    # {speaker2_tag} Нормально, сам как?
-    # {speaker1_tag} Хорошо
-    # {speaker2_tag} Давай поговорим?
-    # {speaker1_tag} О чём хочешь поговорить?"""
-    # print(history)
-    # print('\n[Chat with the model! Send "h" to see the full history]\n')
-    # history = history.split('\n')
-    if sampling_history is not None:
-        with open(sampling_history, 'r') as f:
-            history = f.readlines()
-    else:
-        history = [f"{speaker2_tag} Первое сообщение.\n",
-                   f"{speaker2_tag} Второе сообщение.\n",
-                   f"{speaker2_tag} Третье сообщение.\n",
-                   f"{speaker2_tag} Четвёртое сообщение.\n"]
-    # history = [f"{speaker2_tag} Война на Украине.\n", f"{speaker2_tag} Спецоперация войск РФ.\n",
-    #            f"{speaker2_tag} Путин дал жёсткий ответ западным санкциям.\n",
-    #            f"{speaker2_tag} Зеленский выступил с обращением к нации.\n"]
-    for i in range(max_generation_steps):
-        # message = None #'О стикерах вконтакте'
-        # while not message:
-        #     print(f'{speaker2_tag} writing...:')
-        #     message = input()
-        #     if message == 'h':
-        #         print('\n'.join(history))
-        #         message = None
-        #     elif message == 'quit':
-        #         break
-        # # add new message to history
-        # history.append(f'{speaker2_tag} {message}')
-        # PERSONAL CHAT END
-        # keep only most recent conversation as input to the model
-        recent_history = history[-(2 * max_history):]
-        # concatenate history into single string and add trigger word "bot:"
-        history_str = '{}\n{}'.format('\n'.join(recent_history), speaker1_tag)
-        history_str = '{}\n{}'.format('\n'.join(recent_history), speaker2_tag)
-        # tokenize text and convert into vocabulary ids (input ids)
-        history_enc = tokenizer.encode(history_str, add_special_tokens=True)
-        with torch.no_grad():
-            out_ids = sample_sequence(history_enc, model)
-        out_ids = out_ids[:, len(history_enc):].tolist()[0]
-        if not no_info:
-            print(20 * '-')
-            print('Output of model:')
-            full_output = tokenizer.decode(out_ids, clean_up_tokenization_spaces=True)
-            print(full_output)
-            print('\nInput to the model:')
-            print(history_str)
-            print(20 * '-' + '\n')
-        # Select part before speaker tags as answer
-        i = 0
-        for i, out_id in enumerate(out_ids):
-            if out_id in [speaker1_tag_id, speaker2_tag_id]:
-                break
-        # answer = '{} {}'.format(speaker1_tag, tokenizer.decode(out_ids[:i]))
-        answer = '{} {}'.format(speaker2_tag, tokenizer.decode(out_ids[:i]))
-        print(answer)
-        # add answer to history
-        history.append(answer)
-    return '\n'.join(history)
-
-
-def message_sample(model, tokenizer, sample_file, speaker2_tag='<speaker2>'):
-    with open(sample_file, 'r') as f:
-        beginnings = [' '.join([speaker2_tag, line.strip()]) for line in f.readlines()]
-    history = []
-    for beginning in beginnings:
-        # tokenize text and convert into vocabulary ids (input ids)
-        history_enc = tokenizer.encode(beginning, add_special_tokens=True)
-        with torch.no_grad():
-            out_ids = sample_sequence(history_enc, model)
-        out_ids = out_ids.tolist()[0]
-        answer = tokenizer.decode(out_ids)  #'{} {}'.format(speaker2_tag, tokenizer.decode(out_ids))
-        print(answer)
-        # add answer to history
-        history.append(answer)
-    return '\n'.join(history)
-
-
-def validate(model, data_loader, global_step, writer, device='cuda'):
+def validate(model, tokenizer, data_loader, global_step, writer, device='cuda'):
     model.eval()
     av_loss = 0
     with torch.no_grad():
@@ -144,7 +30,7 @@ def validate(model, data_loader, global_step, writer, device='cuda'):
             inputs, labels = (batch, batch)
             inputs = inputs.to(device)
             labels = labels.to(device)
-            labels[labels == 0] = -100
+            labels[labels == tokenizer.pad_token_id] = -100
             output = model(inputs, labels=labels)
             loss = output['loss']
             val_loss = loss.item()
@@ -173,7 +59,7 @@ def main(cfg):
     tokenizer_class = GPT2Tokenizer  #if "gpt2" in model_type else OpenAIGPTTokenizer
     tokenizer = tokenizer_class.from_pretrained(model_type)
     # Load model
-    model_class = AutoModelWithLMHead  #GPT2LMHeadModel  #if "gpt2" in model_type else OpenAIGPTLMHeadModel
+    model_class = AutoModelWithLMHead
     model = model_class.from_pretrained(model_type)
     model.to(device)
 
@@ -181,9 +67,10 @@ def main(cfg):
 
     # Get data loaders
     logger.info("Prepare datasets")
-    data_loader = get_data_loader(tokenizer, cfg.input_file, train_batch_size=train_batch_size)
+    dataset_mode = 'message' if cfg.message_mode else None
+    data_loader = get_data_loader(tokenizer, cfg.input_file, mode=dataset_mode, batch_size=train_batch_size)
     if cfg.val_file is not None:
-        val_data_loader = get_data_loader(tokenizer, cfg.val_file, shuffle=False, train_batch_size=train_batch_size)
+        val_data_loader = get_data_loader(tokenizer, cfg.val_file, mode=dataset_mode, shuffle=False, batch_size=4 * train_batch_size)
     else:
         val_data_loader = None
     # Prepare optimizer and schedule (linear warmup and decay)
@@ -236,7 +123,7 @@ def main(cfg):
             inputs, labels = (batch, batch)
             inputs = inputs.to(device)
             labels = labels.to(device)
-            labels[labels == 0] = -100  # IGNORE INDEX
+            labels[labels == tokenizer.pad_token_id] = -100  # IGNORE INDEX
             output = model(inputs, labels=labels)
             loss, logits, past_key_values = output['loss'], output['logits'], output['past_key_values']
             loss.backward()
@@ -266,13 +153,16 @@ def main(cfg):
 
                     model.eval()
                     print('*** START SAMPLING ***')
-                    # history = sample(model, tokenizer, cfg.sampling_history)
-                    history = message_sample(model, tokenizer, cfg.sample_file)
+                    if cfg.message_mode:
+                        history = message_sample(model, tokenizer, cfg.sample_file)
+                    else:
+                        history = sample(model, tokenizer, cfg.sampling_history)
                     writer.add_text('Sample', history, global_step=global_step)
+                    print(history)
                     print('*** FINISH SAMPLING ***')
 
                     if val_data_loader is not None:
-                        val_loss = validate(model, val_data_loader, global_step, writer, device)
+                        val_loss = validate(model, tokenizer, val_data_loader, global_step, writer, device)
                         logger.info(f"Current val loss: {val_loss}")
                         if val_loss < best_val_loss:
                             best_val_loss = val_loss
@@ -292,6 +182,31 @@ def main(cfg):
     logger.info(f"Saving model checkpoint to {output_dir}")
     model.save_pretrained(output_dir)
     tokenizer.save_pretrained(output_dir)
+
+
+def parse_args():
+    parser = argparse.ArgumentParser(description='Training params')
+    parser.add_argument('--run_name', default='telegram', type=str)
+    parser.add_argument('--input_file', default='chat.txt', type=str)
+    parser.add_argument('--sample_file', default='chat.txt', type=str)
+    parser.add_argument('--sampling_history', default=None, type=str)
+    parser.add_argument('--message_mode', default=1, type=int)
+    parser.add_argument('--val_file', default=None, type=str)
+    parser.add_argument('--model_type', default='gpt2', type=str)
+    parser.add_argument('--save_every', default=30, type=int)
+    parser.add_argument('--max_input_lenth', default=400, type=int)
+    parser.add_argument('--weight_decay', default=0, type=float)
+    parser.add_argument('--train_batch_size', default=2, type=int)
+    parser.add_argument('--gradient_accumulation_steps', default=16, type=int)
+    parser.add_argument('--warmup_steps', default=8, type=int)
+    parser.add_argument('--lr', default=5e-5, type=float)
+    parser.add_argument('--adam_epsilon', default=1e-8, type=float)
+    parser.add_argument('--max_norm', default=1, type=float)
+    parser.add_argument('--n_epochs', default=12, type=int)
+    parser.add_argument('--device', default='cuda', type=str)
+    parser.add_argument('--seed', default=42, type=int)
+    parser.add_argument('--always_save', default=1, type=int)
+    return parser.parse_args()
 
 
 if __name__ == '__main__':
